# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Forgeflow API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
  /work-items:
    get:
      operationId: listWorkItems
      summary: List or lexically search work items in the selected project
      parameters:
        - { $ref: "#/components/parameters/ProjectIDHeader" }
        - { name: status, in: query, schema: { type: string } }
        - {
            name: type,
            in: query,
            schema: { type: string, enum: [EPIC, STORY, TASK, BUG, SUB_TASK] },
          }
        - {
            name: priority,
            in: query,
            schema:
              { type: string, enum: [LOWEST, LOW, MEDIUM, HIGH, HIGHEST] },
          }
        - {
            name: assignee_id,
            in: query,
            schema: { type: string, format: uuid },
          }
        - { name: sprint_id, in: query, schema: { type: string, format: uuid } }
        - {
            name: repository_id,
            in: query,
            schema: { type: string, format: uuid },
          }
        - { name: q, in: query, schema: { type: string, maxLength: 256 } }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100 },
          }
        - {
            name: cursor,
            in: query,
            schema: { type: string, maxLength: 512 },
            description: Opaque cursor returned by the previous page,
          }
        - {
            name: include_archived,
            in: query,
            schema: { type: boolean, default: false },
            description: Include archived work items in the result,
          }
        - {
            name: sort,
            in: query,
            schema: { type: string, enum: [updated, backlog], default: updated },
            description: Sort by recent activity or persisted backlog rank,
          }
      responses:
        "200":
          description: Work items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    {
                      type: array,
                      items: { $ref: "#/components/schemas/WorkItemResponse" },
                    }
                  next_cursor: { type: string, nullable: true }
    post:
      operationId: createWorkItem
      summary: Create a work item
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkItemRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkItemResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /workspaces:
    get:
      operationId: listWorkspaces
      responses:
        "200":
          description: Workspaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Workspace" }
    post:
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/CreateTenantRequest" } }
      responses: { "201": { description: Workspace created } }
...
  /projects:
    get:
      operationId: listProjects
      parameters:
        - {
            name: workspace_id,
            in: query,
            schema: { type: string, format: uuid },
          }
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Project" }
    post:
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/CreateProjectRequest" } }
      responses: { "201": { description: Project created } }
...
  /sprints:
    get:
      operationId: listSprints
      parameters: [{ $ref: "#/components/parameters/ProjectIDHeader" }]
      responses:
        "200":
          description: Sprints
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Sprint" } }
    post:
      operationId: createSprint
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/CreateSprintRequest" } }
      responses:
        "201":
          description: Sprint created
          content:
            application/json: { schema: { $ref: "#/components/schemas/Sprint" } }
...
  /agent-runs:
    get:
      operationId: listAgentRuns
      parameters:
        - { $ref: "#/components/parameters/ProjectIDHeader" }
        - {
            name: work_item_id,
            in: query,
            schema: { type: string, format: uuid },
          }
      responses: { "200": { description: AgentRuns } }
    post:
      operationId: createAgentRun
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/CreateAgentRunRequest" } }
      responses:
        "201":
          description: AgentRun created
          content:
            application/json: { schema: { $ref: "#/components/schemas/AgentRun" } }
...
  /autonomous-runs:
    get:
      operationId: listAutonomousRuns
      summary: List autonomous workflows in the selected project
      parameters:
        - { $ref: "#/components/parameters/ProjectIDHeader" }
        - { name: work_item_id, in: query, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Autonomous workflows
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AutonomousRun" }
    post:
      operationId: startAutonomousRun
      summary: Start a policy-authorized autonomous workflow
      parameters:
        - { $ref: "#/components/parameters/ProjectIDHeader" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AutonomousStartRequest" }
      responses:
        "201":
          description: Autonomous workflow created
          content:
            application/json: { schema: { $ref: "#/components/schemas/AutonomousRun" } }
...
  /deployments:
    post:
      operationId: requestDeployment
      parameters:
        - { $ref: "#/components/parameters/ProjectIDHeader" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDeploymentRequest" }
      responses:
        "201":
          description: Deployment request created
          content:
            application/json: { schema: { $ref: "#/components/schemas/DeploymentRequest" } }
...
  /audit:
    get:
      operationId: listAuditRecords
      parameters:
        - { name: resource_type, in: query, schema: { type: string } }
        - { name: resource_id, in: query, schema: { type: string } }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100 },
          }
      responses: { "200": { description: Append-only audit records } }
  /mcp:
    post:
      operationId: mcpStreamableHTTP
      summary: Authenticated MCP Streamable HTTP transport
      description: Requires a scoped Bearer PAT and optional X-Project-ID header.
      responses:
        "200": { description: MCP protocol response }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
  /integrations/github/webhooks:
    post:
      operationId: receiveGitHubWebhook
      security: []
      parameters:
        - {
            name: X-GitHub-Delivery,
            in: header,
            required: true,
            schema: { type: string },
          }
        - {
            name: X-GitHub-Event,
            in: header,
            required: true,
            schema: { type: string },
          }
        - {
            name: X-Hub-Signature-256,
            in: header,
            required: true,
            schema: { type: string },
          }
      requestBody:
        required: true
        content:
          application/json:
            { schema: { type: object, additionalProperties: true } }
      responses:
        { "202": { description: Accepted and deduplicated by delivery ID } }
...
  /integrations/github/repositories:
    get:
      operationId: listGitHubRepositories
      summary: Sync and list repositories available to the organization
      parameters:
        - {
            name: project_id,
            in: query,
            required: true,
            schema: { type: string, format: uuid },
          }
      responses:
        "200":
          description: Available GitHub repositories
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    {
                      type: array,
                      items: { $ref: "#/components/schemas/GitHubRepository" },
                    }
...
  /notifications:
    get:
      operationId: listNotifications
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses:
        "200":
          description: Actor-scoped notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Notification" } }
...
components:
  parameters:
...
    ProjectIDHeader:
      name: X-Project-ID
      in: header
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, minLength: 1, maxLength: 200 }
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    CreateWorkItemRequest:
      type: object
      required: [project_id, type, title]
      properties:
        project_id: { type: string, format: uuid }
        type: { type: string, enum: [EPIC, STORY, TASK, BUG, SUB_TASK] }
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 20000 }
        repository_id: { type: string, format: uuid }
        parent_id: { type: string, format: uuid }
        assignee_id: { type: string, format: uuid, nullable: true }
        priority: { type: string, enum: [LOWEST, LOW, MEDIUM, HIGH, HIGHEST] }
        due_at: { type: string, format: date-time, nullable: true }
        estimate_points:
          { type: integer, minimum: 0, maximum: 100, nullable: true }
        sprint_id: { type: string, format: uuid, nullable: true }
...
    WorkItemResponse:
      type: object
      required: [id, key, title, status, version]
      properties:
        id: { type: string, format: uuid }
        key: { type: string }
        title: { type: string }
        description: { type: string }
        status: { type: string }
        version: { type: integer }
        organization_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        type: { type: string }
        parent_id: { type: string, format: uuid, nullable: true }
        priority: { type: string, enum: [LOWEST, LOW, MEDIUM, HIGH, HIGHEST] }
        assignee_id: { type: string, format: uuid, nullable: true }
        reporter_id: { type: string, format: uuid, nullable: true }
        due_at: { type: string, format: date-time, nullable: true }
        estimate_points: { type: integer, minimum: 0, nullable: true }
        backlog_rank: { type: integer, format: int64 }
        repository_id: { type: string, format: uuid, nullable: true }
        sprint_id: { type: string, format: uuid, nullable: true }
        archived_at: { type: string, format: date-time, nullable: true }
        archived_by: { type: string, nullable: true }
...
    Workspace:
      type: object
      required: [id, organization_id, key, display_name, created_at]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        key: { type: string }
        display_name: { type: string }
        created_at: { type: string, format: date-time }
    Project:
      type: object
      required: [id, organization_id, workspace_id, key, display_name, created_at]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        key: { type: string }
        display_name: { type: string }
        created_at: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
        request_id: { type: string }
...
    CreateTenantRequest:
      type: object
      required: [key, display_name]
      properties: { key: { type: string }, display_name: { type: string } }
...
    CreateProjectRequest:
      type: object
      required: [workspace_id, key, display_name]
      properties:
        {
          workspace_id: { type: string, format: uuid },
          key: { type: string },
          display_name: { type: string },
        }
    CreateSprintRequest:
      type: object
      required: [project_id, name]
      properties:
        {
          project_id: { type: string, format: uuid },
          name: { type: string },
          goal: { type: string },
          starts_at: { type: string, format: date, nullable: true },
          ends_at: { type: string, format: date, nullable: true },
        }
    Sprint:
      type: object
      required: [id, organization_id, project_id, name, goal, status, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        name: { type: string }
        goal: { type: string }
        starts_at: { type: string, format: date-time, nullable: true }
        ends_at: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [PLANNED, ACTIVE, COMPLETED] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateAgentRunRequest:
      type: object
      required: [project_id, work_item_id, repository_id, agent_provider, agent_name]
      properties:
        project_id: { type: string, format: uuid }
        work_item_id: { type: string, format: uuid }
        repository_id: { type: string, format: uuid }
        agent_provider: { type: string }
        agent_name: { type: string }
        model: { type: string }
        base_sha: { type: string }
        branch: { type: string }
        execution_inputs: { $ref: "#/components/schemas/AgentRunExecutionInputs" }
        execution_policy: { type: object, additionalProperties: true }
...
    AgentRun:
      type: object
      required: [id, organization_id, project_id, work_item_id, agent_provider, agent_name, status, approval_fingerprint_version, created_at]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        work_item_id: { type: string, format: uuid }
        repository_id: { type: string, format: uuid }
        agent_provider: { type: string }
        agent_name: { type: string }
        model: { type: string }
        base_sha: { type: string }
        branch: { type: string }
        execution_inputs: { $ref: "#/components/schemas/AgentRunExecutionInputs" }
        execution_policy: { type: object, additionalProperties: true }
        approval_fingerprint_version: { type: integer, minimum: 1 }
        approval_fingerprint: { type: string, maxLength: 128 }
        status:
          { $ref: "#/components/schemas/AgentRunStatus" }
        approved: { type: boolean }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
        last_heartbeat_at: { type: string, format: date-time, nullable: true }
        interruption_reason: { type: string }
        commit_sha: { type: string }
        pull_request_id: { type: string, format: uuid }
        result: { type: object, additionalProperties: true }
        error: { type: string }
        metadata: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
...
    CreateDeploymentRequest:
      type: object
      required: [project_id, environment_id, commit_sha]
      properties:
        project_id: { type: string, format: uuid }
        environment_id: { type: string, format: uuid }
        autonomous_run_id: { type: string, format: uuid }
        commit_sha: { type: string, minLength: 1, maxLength: 128 }
...